Added and Tested the Review Manager
I added the review manager which holds the list of reviews and the
getters to get the name of the restaurant the reviews are for, the
reviews themselves, also added methods to add and remove a review, and
get the size of the review. I added the test code needed to ensure
functionality of the review manager class.