Hooked up the reviews functionality to the API
Hooked up the reviews functionality to the API. I also had to edit test
code in order to do so. I edited the review code behind to accept the
view model as a parameter, so that it can inherit the view model
functions. I also allowed for the restaurant view model to accept the
main manager as a parameter so that it can get the main manager's data
for the random restaurant picked.