Created various new tests and modified viewModel and codebehinds
Changed imageProperty to a StringProperty due to inability to properly
test with Images being instantiated in the ViewModel, which required
launching a mock Application, however it is currently beyond my scope to
stop the launched application producing errors when multiple tests
require Application launches. Furthermore, added tests to URLBuilder to
obtain 100% test coverage and modified existing tests to obtain 100%
test coverage. Added preconditions to the SceneController methods as
well.